Update README.md
AssignRef is a platform that allows game official's of the Liga de Fútbol Americano Profesional (LFA) to be assigned games, graded based on their performance, tested on their knowledge, and compensated for their work. AssignRef is an MVP and due to the proprietary nature of the code I can only provide code snippets of my code. This repository contains SQL, .NET, and React code for the test builder feature.